requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
Pillow>=10.0
django-filter>=23.3

# Production
gunicorn>=21.0
whitenoise>=6.6
psycopg2-binary>=2.9
dj-database-url>=2.1
